refactor(spa): extract deleteHostCascade to lib/host-lifecycle (#163)
Move the cascade delete logic from OverviewSection's inline handleDeleteHost
into a reusable deleteHostCascade function. Also adds models snapshot/restore
that was previously missing from the undo path.
